Updated and moved Core and Application headers
