+++ Engine/Source/Runtime/Core/Public/Core.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "Math/Math.h"
+
+#include "Misc/EntryPoint.h"
+#include "Misc/Application.h"
