--- Engine/Source/Runtime/Core/Public/Nexus.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "EntryPoint.h"
-#include "Application.h"
